+++ core/ChangeLog.md
@@ -31,6 +31,7 @@ XP Framework ChangeLog
 
 ### Features
 
+* Added optional "decode" parameter to peer.mail.Message::getBody()  - (@thekid)
 * Added peer.http.FileData class for use in peer.http.FormRequestData. Makes
   it easier to code file uploading - (@thekid)
 * Added possibility for remote.protocol.Serializer to read from input streams